+++ ra2ce/network/network_simplification/__init__.py
@@ -1,3 +1,3 @@
-from ra2ce.network.network_simplification.network_simplification_factory import (
+from ra2ce.network.network_simplification.network_graph_simplificator import (
     NetworkGraphSimplificator,
 )
