+++ ra2ce/network/network_simplification/network_graph_simplificator.py
